pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-devtools"
version = "2.1.0"
authors = [
  { name="daoch4n", email="daoch4n@gmail.com" },
]
description = "multi-functional MCP-devtools server over SSE"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = { text = "Unlicense" }
dependencies = [
    "fastapi",
    "uvicorn",
    "mcp-server-git",
    "httpx",
    "python-dotenv",
"PyYAML",
    
]


[project.scripts]
mcp-devtools = "mcp_devtools_cli:main"

[project.urls]
"Homepage" = "https://github.com/daoch4n/mcp-devtools"
"Bug Tracker" = "https://github.com/daoch4n/mcp-devtools/issues"
[tool.setuptools]
py-modules = ["server", "mcp_devtools_cli"]
